Feat: Pass the number of pages to Pagination
